Added AbstractLogger and ability to log Throwables.
